ref: move agentdetailcontext.tsx
