+++ rnd/autogpt_builder/src/components/marketplace/AgentDetailContent.tsx
